+++ ExilenceClient/src/app/authorize/components/char-profile/char-profile.component.ts
@@ -10,6 +10,8 @@ import { ExternalService } from '../../../shared/providers/external.service';
 import { PartyService } from '../../../shared/providers/party.service';
 import { SessionService } from '../../../shared/providers/session.service';
 import { CharEquipmentComponent } from './char-equipment/char-equipment.component';
+import { Subscription } from 'rxjs/internal/Subscription';
+import { AccountService } from '../../../shared/providers/account.service';
 import { CharMapsComponent } from './char-maps/char-maps.component';
 import { CharWealthComponent } from './char-wealth/char-wealth.component';
 
@@ -20,7 +22,7 @@ import { CharWealthComponent } from './char-wealth/char-wealth.component';
 })
 export class CharProfileComponent implements OnInit, OnDestroy {
   player: Player;
-
+  currentPlayer: Player;
   @Input() localProfile = false;
   @ViewChild('subTabGroup') subTabGroup: MatTabGroup;
   @ViewChild('equipmentTab') equipmentTab: MatTab;
@@ -31,6 +33,7 @@ export class CharProfileComponent implements OnInit, OnDestroy {
   selectedIndex = 0;
   private selectedPlayerSub: Subscription;
   private selectedGenPlayerSub: Subscription;
+  private playerSub: Subscription;
 
   constructor(
     private partyService: PartyService,
@@ -39,6 +42,8 @@ export class CharProfileComponent implements OnInit, OnDestroy {
     private router: Router,
     private electronService: ElectronService,
     private analyticsService: AnalyticsService,
+    private ladderService: LadderService,
+    private accountService: AccountService
   ) {
     this.analyticsService.sendScreenview('/authorized/party/player/profile');
   }
@@ -51,6 +56,9 @@ export class CharProfileComponent implements OnInit, OnDestroy {
           window.dispatchEvent(new Event('resize'));
         }, 1000);
         this.player = res;
+      });
+      this.playerSub = this.accountService.player.subscribe(res => {
+        this.currentPlayer = res;
       });
     } else {
       this.selectedGenPlayerSub = this.partyService.selectedGenericPlayer.subscribe(res => {
@@ -93,6 +101,7 @@ export class CharProfileComponent implements OnInit, OnDestroy {
       case 1: {
         this.charWealth.openCurrencyDialog();
         this.analyticsService.sendScreenview('/authorized/party/player/wealth');
+        this.partyService.updatePlayerGain(this.player, this.currentPlayer.account === this.player.account);
         break;
       }
       // maps