+++ ExilenceClient/src/app/authorize/party/party-summary/party-summary.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PartySummaryComponent } from './party-summary.component';
-import { MatButtonModule, MatInputModule, MatMenuModule, MatIconModule } from '@angular/material';
+import { MatButtonModule, MatInputModule, MatMenuModule, MatIconModule, MatTabsModule, MatFormFieldModule, MatSelectModule } from '@angular/material';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { NetworthTableModule } from '../../components/networth-table/networth-table.module';
 import { IncomeModule } from '../../components/income/income.module';
@@ -20,6 +20,9 @@ import { InfoDialogComponent } from '../../components/info-dialog/info-dialog.co
     MatButtonModule,
     MatMenuModule,
     MatIconModule,
+    MatTabsModule,
+    MatFormFieldModule,
+    MatSelectModule,
     InfoDialogModule
   ],
   declarations: [PartySummaryComponent],