+++ ExilenceClient/src/app/authorize/party/party.component.ts
@@ -12,6 +12,7 @@ import { NetWorthSnapshot } from '../../shared/interfaces/income.interface';
 import { AccountService } from '../../shared/providers/account.service';
 import { PartySummaryComponent } from './party-summary/party-summary.component';
 import { Subscription } from 'rxjs';
+import { SettingsService } from '../../shared/providers/settings.service';
 
 
 @Component({
@@ -30,20 +31,21 @@ export class PartyComponent implements OnInit, OnDestroy {
   private partySub: Subscription;
   private currentPlayerValueSub: Subscription;
   private currentPlayerGainSub: Subscription;
+  private gainHours = 1;
 
   constructor(
     public partyService: PartyService,
     private accountService: AccountService,
     private analyticsService: AnalyticsService,
     private messageValueService: MessageValueService,
-    private electronService: ElectronService
+    private electronService: ElectronService,
+    private settingsService: SettingsService
   ) {
     this.selectedPlayerSub = this.partyService.selectedPlayer.subscribe(res => {
       if (res !== undefined) {
         this.player = res;
         this.messageValueService.playerValue = this.player.netWorthSnapshots[0].value;
-        const isCurrentPlayer = res.account === this.partyService.currentPlayer.account;
-        this.updatePlayerGain(res, false);
+        this.partyService.updatePlayerGain(res, false);
       }
     });
     this.playerSub = this.accountService.player.subscribe(res => {
@@ -58,7 +60,7 @@ export class PartyComponent implements OnInit, OnDestroy {
         // update msg-values based on current player
         this.messageValueService.currentPlayerValueSubject.next(res.netWorthSnapshots[0].value);
         const isCurrentPlayer = res.account === this.partyService.currentPlayer.account;
-        this.updatePlayerGain(res, isCurrentPlayer);
+        this.partyService.updatePlayerGain(res, isCurrentPlayer);
       }
     });
     this.partySub = this.partyService.partyUpdated.subscribe(res => {
@@ -74,6 +76,14 @@ export class PartyComponent implements OnInit, OnDestroy {
         this.messageValueService.partyValueSubject.next(networth);
       }
     });
+    const gainHourSetting = this.settingsService.get('gainHours');
+    if (gainHourSetting !== undefined) {
+      this.gainHours = gainHourSetting;
+    } else {
+      this.gainHours = 1;
+      this.settingsService.set('gainHours', 1);
+    }
+
   }
 
   updatePopout() {
@@ -129,30 +139,4 @@ export class PartyComponent implements OnInit, OnDestroy {
     }
 
   }
-
-  updatePlayerGain(player: Player, current: boolean) {
-    const oneHourAgo = (Date.now() - (1 * 60 * 60 * 1000));
-    const pastHoursSnapshots = player.netWorthSnapshots
-      .filter((snaphot: NetWorthSnapshot) => snaphot.timestamp > oneHourAgo);
-
-    if (pastHoursSnapshots.length > 1) {
-      const lastSnapshot = pastHoursSnapshots[0];
-      const firstSnapshot = pastHoursSnapshots[pastHoursSnapshots.length - 1];
-      const gainHour = ((1000 * 60 * 60)) / (lastSnapshot.timestamp - firstSnapshot.timestamp) * (lastSnapshot.value - firstSnapshot.value);
-      if (current) {
-        this.messageValueService.currentPlayerGainSubject.next(gainHour);
-      } else {
-        this.messageValueService.playerGain = gainHour;
-      }
-    } else {
-      if (current) {
-        this.messageValueService.currentPlayerGainSubject.next(0);
-      } else {
-        this.messageValueService.playerGain = 0;
-      }
-    }
-
-  }
-
-
 }